refactor: features/構造をLaravel 12標準のpages/構造に移行
Laravel 12拡張構成ドキュメントに従い、feature-based構造からpages/構造に移行:

## 移行内容
- **ページコンポーネント**: features/*/pages/* → pages/*/
- **ドメインフック**: features/*/hooks/* → hooks/domain/
- **ドメインダイアログ**: features/*/components/dialogs/* → components/domain/*/dialogs/
- **プロファイルコンポーネント**: features/profile/components/* → components/domain/profile/

## ディレクトリ構造
- pages/Auth/, pages/Book/, pages/BookShelf/, pages/Memo/, pages/Profile/
- components/domain/{book,bookshelf,memo,profile}/dialogs/
- hooks/domain/ (統一されたドメインフック)

## 変更ファイル数
- 移動: 44ファイル
- 更新: インポートパス全体の修正
- 削除: 不要になったfeatures/ディレクトリ
