+++ resources/ts/components/domain/book/card/BookShelfCard.tsx
@@ -1,14 +1,14 @@
 import { BookService } from '@/api/services';
 import { BaseCard } from '@/components/common/BaseCard';
+import { BookDetailDialog } from '@/components/domain/book/dialogs/BookDetailDialog';
+import { UpdateReadStatusDialog } from '@/components/domain/book/dialogs/UpdateReadStatusDialog';
+import { DeleteBookDialog } from '@/components/domain/bookshelf/dialogs/DeleteBookDialog';
+import { CreateMemoDialog } from '@/components/domain/memo/dialogs/CreateMemoDialog';
 import { Button } from '@/components/ui/button';
-import { BookDetailDialog } from '@/features/book/components/dialogs/BookDetailDialog';
-import { UpdateReadStatusDialog } from '@/features/book/components/dialogs/UpdateReadStatusDialog';
-import { useBookMemo } from '@/features/book/hooks/useBookMemo';
-import { useBookShelf } from '@/features/book/hooks/useBookShelf';
-import { useFavoriteBook } from '@/features/book/hooks/useFavoriteBook';
-import { DeleteBookDialog } from '@/features/bookshelf/components/dialogs/DeleteBookDialog';
-import { CreateMemoDialog } from '@/features/memo/components/dialogs/CreateMemoDialog';
 import { useBookCardState } from '@/hooks/domain';
+import { useBookMemo } from '@/hooks/domain/useBookMemo';
+import { useBookShelf } from '@/hooks/domain/useBookShelf';
+import { useFavoriteBook } from '@/hooks/domain/useFavoriteBook';
 import { ReadStatus } from '@/types';
 import type { BookData } from '@/types/api';
 import { Book, BookOpen, Edit, Heart, Library } from 'lucide-react';
