+++ resources/ts/components/domain/book/dialogs/BookDetailDialog.tsx
