+++ resources/ts/components/domain/bookshelf/dialogs/DeleteBookDialog.tsx
