+++ resources/ts/components/domain/bookshelf/dialogs/DeleteBookShelfDialog.tsx
