+++ resources/ts/components/domain/bookshelf/dialogs/ShareLinkDialog.tsx
