+++ resources/ts/components/domain/memo/dialogs/DeleteMemoDialog.tsx
