+++ resources/ts/hooks/domain/useFavoriteBook.ts
