+++ resources/ts/pages/Auth/ConfirmPassword.tsx
