+++ resources/ts/pages/Auth/ForgotPassword.tsx
