+++ resources/ts/pages/Auth/Register.tsx
