+++ resources/ts/pages/Auth/VerifyEmail.tsx
