+++ resources/ts/pages/Book/FavoriteBookList.tsx
