+++ resources/ts/pages/BookShelf/BookShelfDetail.tsx
