+++ resources/ts/pages/BookShelf/BookShelfList.tsx
@@ -2,10 +2,10 @@ import { BookShelfService } from '@/api/services';
 import CommonPagination from '@/components/common/CommonPagination';
 import DefaultLayout from '@/components/common/layout';
 import { BookShelfDefaultCard } from '@/components/domain/bookshelf';
+import { CreateBookShelfDialog } from '@/components/domain/bookshelf/dialogs/CreateBookShelfDialog';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Tabs, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { CreateBookShelfDialog } from '@/features/bookshelf/components/dialogs/CreateBookShelfDialog';
 import type { BookShelfData } from '@/types/api';
 import { Head, router } from '@inertiajs/react';
 import axios from 'axios';
