+++ resources/ts/pages/BookShelf/SharedBookShelfDetail.tsx
