+++ resources/ts/pages/PrivacyPolicy.tsx
