+++ resources/ts/pages/ServerError.tsx
