+++ resources/ts/pages/Welcome.tsx
